# Optimize ENI slot reservation for non-supported instance type