--- a/pkg/ipamd/ipamd_test.go
+++ b/pkg/ipamd/ipamd_test.go
@@ -527,18 +527,21 @@ func getDummyENIMetadataWithV6Prefix() awsutils.ENIMetadata {
 
 func TestIncreaseIPPoolDefault(t *testing.T) {
 	_ = os.Unsetenv(envCustomNetworkCfg)
-	testIncreaseIPPool(t, false, false, false)
+	testIncreaseIPPool(t, false, false, false, false)
+	testIncreaseIPPool(t, false, false, false, true)
 }
 
 func TestIncreaseIPPoolSubnetDiscoveryUnfilledENI(t *testing.T) {
 	_ = os.Unsetenv(envCustomNetworkCfg)
-	testIncreaseIPPool(t, false, false, true)
+	testIncreaseIPPool(t, false, false, true, false)
+	testIncreaseIPPool(t, false, false, true, true)
 }
 
 func TestIncreaseIPPoolCustomENI(t *testing.T) {
 	_ = os.Setenv(envCustomNetworkCfg, "true")
 	_ = os.Setenv("MY_NODE_NAME", myNodeName)
-	testIncreaseIPPool(t, true, false, false)
+	testIncreaseIPPool(t, true, false, false, false)
+	testIncreaseIPPool(t, true, false, false, true)
 }
 
 // Testing that the ENI will be allocated on non schedulable node when the AWS_MANAGE_ENIS_NON_SCHEDULABLE is set to `true`
@@ -546,7 +549,8 @@ func TestIncreaseIPPoolCustomENIOnNonSchedulableNode(t *testing.T) {
 	_ = os.Setenv(envCustomNetworkCfg, "true")
 	_ = os.Setenv(envManageENIsNonSchedulable, "true")
 	_ = os.Setenv("MY_NODE_NAME", myNodeName)
-	testIncreaseIPPool(t, true, true, false)
+	testIncreaseIPPool(t, true, true, false, false)
+	testIncreaseIPPool(t, true, true, false, true)
 }
 
 // Testing that the ENI will NOT be allocated on non schedulable node when the AWS_MANAGE_ENIS_NON_SCHEDULABLE is not set
@@ -554,10 +558,11 @@ func TestIncreaseIPPoolCustomENIOnNonSchedulableNodeDefault(t *testing.T) {
 	_ = os.Unsetenv(envManageENIsNonSchedulable)
 	_ = os.Setenv(envCustomNetworkCfg, "true")
 	_ = os.Setenv("MY_NODE_NAME", myNodeName)
-	testIncreaseIPPool(t, true, true, false)
+	testIncreaseIPPool(t, true, true, false, false)
+	testIncreaseIPPool(t, true, true, false, true)
 }
 
-func testIncreaseIPPool(t *testing.T, useENIConfig bool, unschedulabeNode bool, subnetDiscovery bool) {
+func testIncreaseIPPool(t *testing.T, useENIConfig bool, unschedulableNode bool, subnetDiscovery bool, eniTrunking bool) {
 	m := setup(t)
 	defer m.ctrl.Finish()
 	ctx := context.Background()
@@ -632,17 +637,17 @@ func testIncreaseIPPool(t *testing.T, useENIConfig bool, unschedulabeNode bool, 
 		},
 	}
 
-	if unschedulabeNode {
+	if unschedulableNode {
 		val, exist := os.LookupEnv(envManageENIsNonSchedulable)
 		if exist && val == "true" {
-			assertAllocationExternalCalls(true, useENIConfig, m, sg, podENIConfig, eni2, eniMetadata, false)
+			assertAllocationExternalCalls(true, useENIConfig, m, sg, podENIConfig, eni2, eniMetadata, false, eniTrunking)
 		} else {
-			assertAllocationExternalCalls(false, useENIConfig, m, sg, podENIConfig, eni2, eniMetadata, false)
+			assertAllocationExternalCalls(false, useENIConfig, m, sg, podENIConfig, eni2, eniMetadata, false, eniTrunking)
 		}
 	} else if subnetDiscovery {
-		assertAllocationExternalCalls(true, useENIConfig, m, sg, podENIConfig, eni2, eniMetadata, true)
+		assertAllocationExternalCalls(true, useENIConfig, m, sg, podENIConfig, eni2, eniMetadata, true, eniTrunking)
 	} else {
-		assertAllocationExternalCalls(true, useENIConfig, m, sg, podENIConfig, eni2, eniMetadata, false)
+		assertAllocationExternalCalls(true, useENIConfig, m, sg, podENIConfig, eni2, eniMetadata, false, eniTrunking)
 	}
 
 	if mockContext.useCustomNetworking {
@@ -658,7 +663,7 @@ func testIncreaseIPPool(t *testing.T, useENIConfig bool, unschedulabeNode bool, 
 			Spec:       v1.NodeSpec{},
 			Status:     v1.NodeStatus{},
 		}
-		if unschedulabeNode {
+		if unschedulableNode {
 			fakeNode.Spec.Taints = append(fakeNode.Spec.Taints, corev1.Taint{
 				Key:    "node.kubernetes.io/unschedulable",
 				Effect: corev1.TaintEffectNoSchedule,
@@ -681,13 +686,15 @@ func testIncreaseIPPool(t *testing.T, useENIConfig bool, unschedulabeNode bool, 
 	mockContext.increaseDatastorePool(ctx, defaultNetworkCard)
 }
 
-func assertAllocationExternalCalls(shouldCall bool, useENIConfig bool, m *testMocks, sg []*string, podENIConfig *eniconfigscheme.ENIConfigSpec, eni2 string, eniMetadata []awsutils.ENIMetadata, subnetDiscovery bool) {
+func assertAllocationExternalCalls(shouldCall bool, useENIConfig bool, m *testMocks, sg []*string, podENIConfig *eniconfigscheme.ENIConfigSpec, eni2 string, eniMetadata []awsutils.ENIMetadata, subnetDiscovery bool, eniTrunking bool) {
 	callCount := 0
 	if shouldCall {
 		callCount = 1
 	}
 
 	originalErr := errors.New("err")
+
+	m.awsutils.EXPECT().IsTrunkingCompatible().Times(callCount).Return(eniTrunking)
 
 	if useENIConfig {
 		m.awsutils.EXPECT().AllocENI(sg, podENIConfig.Subnet, 14, 0).Times(callCount).Return(eni2, nil)
@@ -715,20 +722,26 @@ func assertAllocationExternalCalls(shouldCall bool, useENIConfig bool, m *testMo
 
 func TestIncreasePrefixPoolDefault(t *testing.T) {
 	_ = os.Unsetenv(envCustomNetworkCfg)
-	testIncreasePrefixPool(t, false, false)
+	_ = os.Setenv("MY_NODE_NAME", myNodeName)
+	testIncreasePrefixPool(t, false, false, true)
+	testIncreasePrefixPool(t, false, false, false)
 }
 
 func TestIncreasePrefixPoolSubnetDiscoveryUnfilledENI(t *testing.T) {
 	_ = os.Unsetenv(envCustomNetworkCfg)
-	testIncreasePrefixPool(t, false, true)
+	_ = os.Setenv("MY_NODE_NAME", myNodeName)
+	testIncreasePrefixPool(t, false, true, true)
+	testIncreasePrefixPool(t, false, true, false)
 }
 
 func TestIncreasePrefixPoolCustomENI(t *testing.T) {
 	_ = os.Setenv(envCustomNetworkCfg, "true")
-	testIncreasePrefixPool(t, true, false)
+	_ = os.Setenv("MY_NODE_NAME", myNodeName)
+	testIncreasePrefixPool(t, true, false, true)
+	testIncreasePrefixPool(t, true, false, false)
 }
 
-func testIncreasePrefixPool(t *testing.T, useENIConfig, subnetDiscovery bool) {
+func testIncreasePrefixPool(t *testing.T, useENIConfig, subnetDiscovery bool, eniTrunking bool) {
 	m := setup(t)
 	defer m.ctrl.Finish()
 	ctx := context.Background()
@@ -774,6 +787,8 @@ func testIncreasePrefixPool(t *testing.T, useENIConfig, subnetDiscovery bool) {
 	}
 
 	originalErr := errors.New("err")
+
+	m.awsutils.EXPECT().IsTrunkingCompatible().Return(eniTrunking)
 
 	if useENIConfig {
 		m.awsutils.EXPECT().AllocENI(sg, podENIConfig.Subnet, 1, defaultNetworkCard).Return(eni2, nil)
@@ -948,6 +963,7 @@ func TestTryAddIPToENI(t *testing.T) {
 
 	mockContext.dataStoreAccess = testDatastore()
 
+	m.awsutils.EXPECT().IsTrunkingCompatible().Return(true)
 	m.awsutils.EXPECT().AllocENI(nil, "", warmIPTarget, defaultNetworkCard).Return(secENIid, nil)
 	eniMetadata := []awsutils.ENIMetadata{
 		{