**What type of PR is this?**
improvement

**Which issue does this PR fix?**:
Close #3094

**What does this PR do / Why do we need it?**:
- Do not reserve ENI slot for non-supported instance type
- Integration test util add `CreateAndWaitTillManagedNGReady`

**Testing done on this change**:
`make unit-test`

### Integration test
```console
# ginkgo -v --fail-on-pending --  --cluster-kubeconfig=$KUBECONFIG  --cluster-name=$CLUSTER_NAME  --aws-region=$AWS_REGION  --aws-vpc-id=$VPC_ID
2025/09/23 22:03:22 maxprocs: Leaving GOMAXPROCS=8: CPU quota undefined
Running Suite: ENI Trunking Suite - /home/hungtran/dev/amazon-vpc-cni-k8s/test/integration/eni-trunking
=======================================================================================================
Random Seed: 1758639804

Will run 1 of 1 specs
------------------------------
[BeforeSuite]
/home/hungtran/dev/amazon-vpc-cni-k8s/test/integration/eni-trunking/eni_trunking_suite_test.go:31
  STEP: creating test namespace @ 09/23/25 22:03:40.419
  STEP: Getting Private subnets @ 09/23/25 22:03:41.199
  STEP: Deploying non-eni-trunking t3.medium managed nodegroup of size 1 @ 09/23/25 22:03:44.803
[controller-runtime] log.SetLogger(...) was never called; logs will not be displayed.
Detected at:
        >  goroutine 198 [running]:
        >  runtime/debug.Stack()
        >       /snap/go/10938/src/runtime/debug/stack.go:26 +0x5e
        >  sigs.k8s.io/controller-runtime/pkg/log.eventuallyFulfillRoot()
        >       /root/go/pkg/mod/sigs.k8s.io/controller-runtime@v0.21.0/pkg/log/log.go:60 +0xcd
        >  sigs.k8s.io/controller-runtime/pkg/log.(*delegatingLogSink).Enabled(0xc0007a2600, 0x3)
        >       /root/go/pkg/mod/sigs.k8s.io/controller-runtime@v0.21.0/pkg/log/deleg.go:111 +0x32
        >  github.com/go-logr/logr.Logger.Info({{0x6a4ca00?, 0xc0007a2600?}, 0xc000974000?}, {0x633272e, 0x12}, {0xc000d44c60, 0x6, 0x6})
        >       /root/go/pkg/mod/github.com/go-logr/logr@v1.4.2/logr.go:276 +0x6e
        >  k8s.io/client-go/tools/cache.(*Reflector).RunWithContext(0xc000256780, {0x6a44b98, 0xc0007662d0})
        >       /root/go/pkg/mod/k8s.io/client-go@v0.33.4/tools/cache/reflector.go:357 +0x1c8
        >  k8s.io/client-go/tools/cache.(*controller).RunWithContext.(*Group).StartWithContext.func3()
        >       /root/go/pkg/mod/k8s.io/apimachinery@v0.33.4/pkg/util/wait/wait.go:63 +0x1f
        >  k8s.io/apimachinery/pkg/util/wait.(*Group).Start.func1()
        >       /root/go/pkg/mod/k8s.io/apimachinery@v0.33.4/pkg/util/wait/wait.go:72 +0x4c
        >  created by k8s.io/apimachinery/pkg/util/wait.(*Group).Start in goroutine 176
        >       /root/go/pkg/mod/k8s.io/apimachinery@v0.33.4/pkg/util/wait/wait.go:70 +0x73
[BeforeSuite] PASSED [131.338 seconds]
------------------------------
ENI Trunking Suite ENABLE_POD_ENI=true Non ENI trunking instance can scale to maxPods
/home/hungtran/dev/amazon-vpc-cni-k8s/test/integration/eni-trunking/eni_trunking_test.go:23
  STEP: setting the environment variables on the ds to map[ENABLE_POD_ENI:true] @ 09/23/25 22:06:02.66
  STEP: getting the aws-node daemon set in namespace kube-system @ 09/23/25 22:06:02.66
  STEP: updating the daemon set with new environment variable @ 09/23/25 22:06:02.961
  STEP: Deploying 13 Busybox pods @ 09/23/25 22:06:09.248
• [25.472 seconds]
------------------------------
[AfterSuite]
/home/hungtran/dev/amazon-vpc-cni-k8s/test/integration/eni-trunking/eni_trunking_suite_test.go:61
  STEP: Deleting test namespace @ 09/23/25 22:06:28.132
  STEP: Deleting Managed Nodegroup @ 09/23/25 22:06:44.33
[AfterSuite] PASSED [395.832 seconds]
------------------------------

Ran 1 of 1 Specs in 552.642 seconds
SUCCESS! -- 1 Passed | 0 Failed | 0 Pending | 0 Skipped
PASS

Ginkgo ran 1 suite in 9m22.027200733s
Test Suite Passed
```

### Manual test
#### Prerequisites:
SGP enabled
<img width="613" height="100" alt="Screenshot 2025-09-06 130136" src="https://github.com/user-attachments/assets/7c4b438d-275c-42b5-9770-d7509280ee01" />

Instance types for testing:
- t3.medium, no trunking, maxPods=17
<img width="562" height="82" alt="Screenshot 2025-09-06 125251" src="https://github.com/user-attachments/assets/07574283-aa24-4c4e-aaa5-67b2f3c72056" />

- m7a.medium, trunking supported, maxPods=8
<img width="556" height="98" alt="Screenshot 2025-09-06 125214" src="https://github.com/user-attachments/assets/203705ff-c591-42a9-bd2b-31ca01e7f811" />

#### Reproducing
On latest CNI, no-trunking cannot scale to maxPods
<img width="929" height="266" alt="Screenshot 2025-09-06 130044" src="https://github.com/user-attachments/assets/38dd02e5-b76b-4f66-9aaf-22ca7c35da35" />

#### Test new code
1. Build test image `IMAGE=phuhung273/amazon-k8s-cni make docker`: phuhung273/amazon-k8s-cni:089fc089
2. Modify aws-node to use test image
3. no-trunking can scale to maxPods
<img width="917" height="274" alt="Screenshot 2025-09-06 124920" src="https://github.com/user-attachments/assets/1250edb1-c101-42eb-a2c4-90fb82b85a9e" />

4. trunking still cannot scale to maxPods
<img width="942" height="176" alt="Screenshot 2025-09-06 124834" src="https://github.com/user-attachments/assets/dc42a0bb-147f-4db0-9a2a-1729f93dacd4" />

ipamd.log
```console
{"level":"info","ts":"2025-09-06T05:40:19.098Z","caller":"ipamd/ipamd.go:775","msg":"Successfully added feature SecurityGroupsForPods to CNINode if not existing"}
{"level":"debug","ts":"2025-09-06T05:40:19.098Z","caller":"ipamd/ipamd.go:779","msg":"IP pool is too low for Network Card 0: available (0) < ENI target (1) * addrsPerENI (3)"}
{"level":"debug","ts":"2025-09-06T05:40:19.098Z","caller":"ipamd/ipamd.go:2523","msg":"IP pool stats for network card 0: Total IPs/Prefixes = 3/0, AssignedIPs/CooldownIPs: 3/0, c.maxIPsPerENI = 3"}
{"level":"debug","ts":"2025-09-06T05:40:19.098Z","caller":"ipamd/ipamd.go:765","msg":"IP stats for Network Card 0 - total IPs: 3, assigned IPs: 3, cooldown IPs: 0"}
{"level":"debug","ts":"2025-09-06T05:40:19.098Z","caller":"ipamd/ipamd.go:783","msg":"Starting to increase pool size for network card 0"}
{"level":"debug","ts":"2025-09-06T05:40:19.098Z","caller":"ipamd/ipamd.go:936","msg":"Node found \"ip-172-31-6-102.ec2.internal\" - no of taints - 1"}
{"level":"debug","ts":"2025-09-06T05:40:19.098Z","caller":"ipamd/ipamd.go:783","msg":"Skipping ENI allocation as the max ENI limit is already reached"}
```

**Will this PR introduce any new dependencies?**:
No

**Will this break upgrades or downgrades? Has updating a running cluster been tested?**:
Can the team show me how to test this case ?

**Does this change require updates to the CNI daemonset config files to work?**:
No

**Does this PR introduce any user-facing change?**:
<!--
If yes, a release note update is required:
Enter your extended release note in the block below. If the PR requires additional actions
from users switching to the new release, include the string "action required".
-->

```release-note
Optimize ENI slot reservation for non-supported instance type
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.