@jaydeokar 2025-09-04
Thanks for making this change. Since the instance types and families which currently don't support trunking is fixed, lets add this to the scripts/gen_vpc_ip_limits.go and add a property IsTrunkingEnabled. 
This script generates the file pkg/vpc/vpc_ip_resource_limit.go which CNI refers for getting instance details.

The default value should be true, but only set false for the instance types and families which you've added in the new file