@jaydeokar 2025-09-17
Yes please could you add a small test suite ? You can create managed nodegroups as well, doesn't have to be self managed in this case. Just need to pass the right instance type. 

The change overall looks good to me.. Thanks for working on this